Added weather images
WeatherCollectionViewCell configured UI

WeatherViewController added collectionView

+++ WeatherForecast/WeatherForecast/Models/RLMModels/RLMWeather.swift
@@ -14,10 +14,10 @@ class RLMWeather: RLMModel {
     @objc dynamic var main: String?
     @objc dynamic var info: String?
     @objc dynamic var icon: String?
-    let tempDay = RealmOptional<Int>()
-    let tempMin = RealmOptional<Int>()
-    let tempMax = RealmOptional<Int>()
-    let tempNight = RealmOptional<Int>()
-    let tempEve = RealmOptional<Int>()
-    let tempMorn = RealmOptional<Int>()
+    let tempDay = RealmOptional<Double>()
+    let tempMin = RealmOptional<Double>()
+    let tempMax = RealmOptional<Double>()
+    let tempNight = RealmOptional<Double>()
+    let tempEve = RealmOptional<Double>()
+    let tempMorn = RealmOptional<Double>()
 }
